Swapped update_gui function of the main ui to recieve UICommands instead of strings. Moved SkeletonUI to user_interface folder
