+++ src/rattlesnake/environment/abstract_environment.py
@@ -35,6 +35,7 @@ from rattlesnake.utilities import (
     GlobalCommands,
     VerboseMessageQueue,
 )
+from rattlesnake.user_interface.ui_utilities import UICommands
 
 PICKLE_ON_ERROR = False
 
@@ -320,7 +321,7 @@ class AbstractEnvironment(ABC):
                 self.log(f"ERROR\n\n {tb}")
                 self.gui_update_queue.put(
                     (
-                        "error",
+                        UICommands.ERROR,
                         (
                             f"{self.environment_name} Error",
                             f"!!!UNKNOWN ERROR!!!\n\n{tb}",
